+++ src/api/v1/profile/profile.constant.ts
@@ -1,4 +1,5 @@
 export const success = {
   PROFILE_UPDATED_SUCCESSFULLY: "Profile updated successfully",
   PROFILE_FETCHED_SUCCESSFULLY: "Profile fetched successfully",
+  PROFILE_VISIBILITY_UPDATED_SUCCESSFULLY: "Profile visibility updated successfully",
 };