+++ src/api/v1/profile/profile.dal.ts
@@ -4,6 +4,7 @@ import { ProfileData } from "./profile.validation";
 interface IProfileDataDal {
   upSertProfileData(data: ProfileData): Promise<ProfileData>;
   getProfileData(userId: string): Promise<ProfileData | null>;
+  toggleProfileVisibility(userId: string, isPublic: boolean): Promise<ProfileData | null>;
 }
 
 export class ProfileDataDAL implements IProfileDataDal {
@@ -16,5 +17,9 @@ export class ProfileDataDAL implements IProfileDataDal {
 
   async getProfileData(userId: string): Promise<ProfileData | null> {
     return await ProfileDataModel.findOne({ userId });
+  }
+
+  async toggleProfileVisibility(userId: string, isPublic: boolean): Promise<ProfileData | null> {
+    return await ProfileDataModel.findOneAndUpdate({ userId }, { isPublic }, { new: true });
   }
 }