+++ src/api/v1/profile/profile.route.ts
@@ -8,5 +8,6 @@ const router = Router();
 
 router.put("/profile", validateAccessToken, ProfileController.upSertProfile);
 router.get("/profile", validateAccessToken, ProfileController.getProfile);
+router.patch("/profile/visibility", validateAccessToken, ProfileController.toggleProfileVisibility);
 
 export { router as profileRoutes };